ai-services/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
torch==2.1.1
torchvision==0.16.1
torchaudio==2.1.1
transformers==4.35.2
accelerate==0.24.1
openai-whisper==20231117
ollama==0.1.7
python-dotenv==1.0.0